Updated settings and added some more validation
