+++ windows-terminal-quake/Settings/AnimationType.cs
@@ -1,0 +1,15 @@
+﻿namespace WindowsTerminalQuake.Settings
+{
+	public enum AnimationType
+	{
+		Linear = 0,
+
+		EaseInBack,
+		EaseInCubic,
+		EaseInOutSine,
+		EaseInQuart,
+		EaseOutBack,
+		EaseOutCubic,
+		EaseOutQuart,
+	}
+}
