+++ windows-terminal-quake/Settings/PreferMonitor.cs
@@ -1,0 +1,10 @@
+﻿namespace WindowsTerminalQuake.Settings
+{
+	public enum PreferMonitor
+	{
+		WithCursor = 0,
+
+		AtIndex,
+		Primary,
+	}
+}
